requirements.txt
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytesseract==0.3.10
opencv-python==4.8.0.76
pillow==10.0.0 